backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LLM Integration
google-generativeai==0.3.2
tiktoken==0.5.2

# Vector Database
faiss-cpu==1.7.4
numpy==1.26.3

# Graph Database
neo4j==5.16.0

# Code Analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-typescript==0.20.5

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
